Added Padle an Excluder

+++ Utils/Utils.hpp
@@ -12,5 +12,5 @@ const float TWO_PI = 2.0f * PI;
 bool IsEqual(float x, float y);
 bool IsGreaterThanOrEqual(float x, float y);
 bool IsLessThanOrEqual(float x, float y);
-float MillisecondsToSeconds(float milliseconds);
+float MillisecondsToSeconds(unsigned int milliseconds);
 #endif
